+++ src/js_modules/syllabus/TimelineSidebar.jsx
@@ -1,0 +1,149 @@
+import { ChevronLeftIcon, ChevronRightIcon } from '@chakra-ui/icons';
+import {
+  Box, Checkbox, IconButton,
+} from '@chakra-ui/react';
+import PropTypes from 'prop-types';
+import useTranslation from 'next-translate/useTranslation';
+import Heading from '../../common/components/Heading';
+import { Config, getSlideProps } from './config';
+import Timeline from '../../common/components/Timeline';
+
+const TimelineSidebar = ({
+  cohortSession, filterEmptyModules, onClickAssignment, showPendingTasks, setShowPendingTasks,
+  isOpen, onToggle,
+}) => {
+  const { t } = useTranslation('syllabus');
+  const Open = !isOpen;
+  const slide = getSlideProps(Open);
+  const {
+    themeColor, commonBorderColor, currentThemeValue, colorLight,
+  } = Config();
+
+  return (
+    <>
+      <IconButton
+        style={{ zIndex: 20 }}
+        aria-label="Open Timeline"
+        variant="default"
+        display={Open ? 'none' : 'initial'}
+        onClick={onToggle}
+        width="17px"
+        height="36px"
+        minW={0}
+        position="fixed"
+        top="50%"
+        left="0"
+        padding={0}
+        icon={(
+          <ChevronRightIcon
+            width="17px"
+            height="36px"
+          />
+        )}
+      />
+      <Box position={{ base: 'fixed', lg: Open ? 'initial' : 'fixed' }} display={Open ? 'initial' : 'none'} flex="0 0 auto" minWidth="290px" width={{ base: '74.6vw', md: '46.6vw', lg: '26.6vw' }} zIndex={Open ? 99 : 0}>
+        <Box style={slide}>
+          <Box
+            padding="1.5rem"
+            // position="sticky"
+            display="flex"
+            flexDirection="column"
+            gridGap="6px"
+            top={0}
+            zIndex={200}
+            bg={themeColor}
+            borderBottom={1}
+            borderStyle="solid"
+            borderColor={commonBorderColor}
+          >
+            {cohortSession?.syllabus_version && (
+              <Heading size="xsm">{cohortSession?.syllabus_version?.name}</Heading>
+            )}
+            <Checkbox mb="-14px" onChange={(e) => setShowPendingTasks(e.target.checked)} color={colorLight}>
+              {t('dashboard:modules.show-pending-tasks')}
+            </Checkbox>
+          </Box>
+
+          <IconButton
+            aria-label="Close Timeline"
+            style={{ zIndex: 20 }}
+            variant="default"
+            onClick={onToggle}
+            width="17px"
+            height="36px"
+            minW={0}
+            position="absolute"
+            transition={Open ? 'margin 225ms cubic-bezier(0, 0, 0.2, 1) 0ms' : 'margin 195ms cubic-bezier(0.4, 0, 0.6, 1) 0ms'}
+            transitionProperty="margin"
+            transitionDuration={Open ? '225ms' : '195ms'}
+            transitionTimingFunction={Open ? 'cubic-bezier(0, 0, 0.2, 1)' : 'cubic-bezier(0.4, 0, 0.6, 1)'}
+            top="50%"
+            right="-20px"
+            padding={0}
+            icon={(
+              <ChevronLeftIcon
+                width="17px"
+                height="36px"
+              />
+            )}
+            marginBottom="1rem"
+          />
+
+          <Box
+            className={`horizontal-sroll ${currentThemeValue}`}
+            height="100%"
+            style={{
+              overflowX: 'hidden',
+              overflowY: 'auto',
+            }}
+          >
+            {filterEmptyModules.length > 0 && filterEmptyModules.map((section) => {
+              const currentAssignments = showPendingTasks
+                ? section.filteredModulesByPending
+                : section.filteredModules;
+              return (
+                <Box
+                  key={`${section.title}-${section.id}`}
+                  padding={{ base: '1rem', md: '1.5rem' }}
+                  borderBottom={1}
+                  borderStyle="solid"
+                  borderColor={commonBorderColor}
+                >
+                  <Timeline
+                    key={section.id}
+                    showPendingTasks={showPendingTasks}
+                    assignments={currentAssignments}
+                    technologies={section.technologies || []}
+                    title={section.label}
+                    onClickAssignment={onClickAssignment}
+                  />
+                </Box>
+              );
+            })}
+          </Box>
+        </Box>
+      </Box>
+    </>
+  );
+};
+
+TimelineSidebar.propTypes = {
+  cohortSession: PropTypes.objectOf(PropTypes.any),
+  filterEmptyModules: PropTypes.arrayOf(PropTypes.any),
+  onClickAssignment: PropTypes.func,
+  showPendingTasks: PropTypes.bool,
+  isOpen: PropTypes.bool,
+  setShowPendingTasks: PropTypes.func,
+  onToggle: PropTypes.func,
+};
+TimelineSidebar.defaultProps = {
+  cohortSession: null,
+  filterEmptyModules: [],
+  onClickAssignment: () => {},
+  showPendingTasks: false,
+  isOpen: false,
+  setShowPendingTasks: () => {},
+  onToggle: () => {},
+};
+
+export default TimelineSidebar;
