+++ styles/markdown.css
@@ -480,7 +480,7 @@ blockquote > :last-child {
   margin: 0 2px;
   padding: 2px 5px;
   white-space: nowrap;
-  background-color: #17202A;
+  background-color: #28292A;
   color: #ccc;
   /* border-radius: 3px; */
 }
